requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
requests>=2.31.0
instaloader>=4.10
yt-dlp>=2024.0
flask>=3.0.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
apscheduler>=3.10.0
gunicorn>=21.0.0
